confirm retroactive placement edits
